Fix: Updating doc strings per review, moving files
