+++ telegram/__init__.py
@@ -286,12 +286,7 @@ from ._inline.inputtextmessagecontent import InputTextMessageContent
 from ._inline.inputvenuemessagecontent import InputVenueMessageContent
 from ._keyboardbutton import KeyboardButton
 from ._keyboardbuttonpolltype import KeyboardButtonPollType
-from ._keyboardbuttonrequest import (
-    ChatShared,
-    KeyboardButtonRequestChat,
-    KeyboardButtonRequestUser,
-    UserShared,
-)
+from ._keyboardbuttonrequest import KeyboardButtonRequestChat, KeyboardButtonRequestUser
 from ._loginurl import LoginUrl
 from ._menubutton import MenuButton, MenuButtonCommands, MenuButtonDefault, MenuButtonWebApp
 from ._message import Message
@@ -335,6 +330,7 @@ from ._proximityalerttriggered import ProximityAlertTriggered
 from ._replykeyboardmarkup import ReplyKeyboardMarkup
 from ._replykeyboardremove import ReplyKeyboardRemove
 from ._sentwebappmessage import SentWebAppMessage
+from ._shared import ChatShared, UserShared
 from ._telegramobject import TelegramObject
 from ._update import Update
 from ._user import User
